packaged-food-rating-app/app/requirements.txt
streamlit==1.27.0
pillow>=7.1.0,<10.0.0
pytesseract==0.3.10
requests==2.31.0
numpy>=1.19.3,<2
pandas>=1.3.0,<3
opencv-python>=4.8.0
pyzbar>=0.1.9
imutils>=0.5.4